Структура CurrentAuthUser переделана в класс Функционал класса расширен Клас выполняет загрузку фото на сервер и удаление самостоятельно
